Cellula_3week_Omar_Arnous/requirements.txt
streamlit
faiss-cpu # faiss-gpu if you have a compatible GPU
numpy
huggingface-hub
openai
python-dotenv
sentence-transformers
datasets
transformers
dotenv